requirements.txt
Flask==2.3.2
Flask-Cors==3.1.0
web3==6.4.0
py-solc-x==1.1.1
